refactor: move HTML report to plugins/, lazy-load on demand
