+++ tests/perf/plugins/html_report.py
